Add super trait Typed2718 to Encodable2718
