+++ crates/consensus/src/transaction/pooled.rs
@@ -4,12 +4,12 @@
 use crate::{
     transaction::{RlpEcdsaTx, TxEip1559, TxEip2930, TxEip4844, TxLegacy},
     SignableTransaction, Signed, Transaction, TxEip4844WithSidecar, TxEip7702, TxEnvelope, TxType,
-    Typed2718,
 };
 use alloy_eips::{
     eip2718::{Decodable2718, Eip2718Error, Eip2718Result, Encodable2718},
     eip2930::AccessList,
     eip7702::SignedAuthorization,
+    Typed2718,
 };
 use alloy_primitives::{
     bytes, Bytes, ChainId, PrimitiveSignature as Signature, TxHash, TxKind, B256, U256,
@@ -79,7 +79,7 @@ impl PooledTransaction {
     /// length of the header + the length of the type flag and inner encoding.
     fn network_len(&self) -> usize {
         let mut payload_length = self.encode_2718_len();
-        if !Encodable2718::is_legacy(self) {
+        if !self.is_legacy() {
             payload_length += Header { list: false, payload_length }.length();
         }
 
@@ -282,16 +282,6 @@ impl Decodable for PooledTransaction {
 }
 
 impl Encodable2718 for PooledTransaction {
-    fn type_flag(&self) -> Option<u8> {
-        match self {
-            Self::Legacy(_) => None,
-            Self::Eip2930(_) => Some(0x01),
-            Self::Eip1559(_) => Some(0x02),
-            Self::Eip4844(_) => Some(0x03),
-            Self::Eip7702(_) => Some(0x04),
-        }
-    }
-
     fn encode_2718_len(&self) -> usize {
         match self {
             Self::Legacy(tx) => tx.eip2718_encoded_length(),
