+++ crates/eips/Cargo.toml
@@ -54,6 +54,9 @@ ethereum_ssz = { workspace = true, optional = true }
 # arbitrary
 arbitrary = { workspace = true, features = ["derive"], optional = true }
 
+# misc
+auto_impl.workspace = true
+
 [dev-dependencies]
 alloy-primitives = { workspace = true, features = [
     "rand",
